+++ src/libawkward/array/NumpyArray.cpp
@@ -589,7 +589,7 @@ namespace awkward {
       (int64_t)shape_[1],
       len);
     util::handle_error(err, classname(), identities_.get());
-    std::vector<ssize_t> shape({ len });
+    std::vector<ssize_t> shape({ (ssize_t)len });
     std::vector<ssize_t> strides({ sizeof(int64_t) });
 #ifdef _MSC_VER
     std::string format = "q";
